Resolve CodeQL review blockers
